requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
idna==3.7
multidict==6.0.5
public-ip==0.12
python-weather==2.0.3
ratelim==0.1.6
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
